requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-recaptcha==3.0.0
idna==3.4
python-dateutil==2.8.2
python-decouple==3.8
requests==2.31.0
sib-api-v3-sdk==7.6.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.3
